Add --disable-caching flag
This flag allows the end user to turn off the caching, meaning every
file will be analyzed by an llm, regardless as to whether or not a
cache folder exists already.

Also cleaned up the code a little
